wrap stake page
